Move VC-API exchange processing to `vcapi.js` file.
